# docs/swagger/swagger.yaml
basePath: /
definitions:
  services.ChangeAccountForm:
    properties:
      newUsername:
        maxLength: 256
        type: string
    required:
    - newUsername
    type: object
  services.ChangePasswordForm:
    properties:
      newPassword:
        type: string
      oldPassword:
        type: string
    required:
    - newPassword
    - oldPassword
    type: object
  services.CompleteChangeAccountForm:
    properties:
      code:
        type: string
      newUsername:
        maxLength: 256
        type: string
    required:
    - code
    - newUsername
    type: object
  services.CompleteResetPasswordForm:
    properties:
      code:
        type: string
      newPassword:
        type: string
      username:
        maxLength: 256
        type: string
      validateOnly:
        type: boolean
    required:
    - code
    - newPassword
    - username
    type: object
  services.CompleteVerifyAccountForm:
    properties:
      code:
        type: string
      username:
        maxLength: 256
        type: string
    required:
    - code
    - username
    type: object
  services.CreateAccountForm:
    properties:
      firstName:
        maxLength: 256
        type: string
      lastName:
        maxLength: 256
        type: string
      password:
        type: string
      username:
        maxLength: 256
        type: string
      validateOnly:
        type: boolean
    required:
    - firstName
    - lastName
    - password
    - username
    type: object
  services.ResetPasswordForm:
    properties:
      username:
        maxLength: 256
        type: string
    required:
    - username
    type: object
  services.SignInForm:
    properties:
      password:
        type: string
      username:
        maxLength: 256
        type: string
    required:
    - password
    - username
    type: object
  services.SignInWithRefreshTokenForm:
    properties:
      refreshToken:
        type: string
    required:
    - refreshToken
    type: object
  services.SignOutForm:
    properties:
      global:
        type: boolean
      refreshToken:
        type: string
    required:
    - refreshToken
    type: object
  services.VerifyAccountForm:
    properties:
      username:
        maxLength: 256
        type: string
    required:
    - username
    type: object
info:
  contact: {}
  description: Konabra is a smart, community-powered transport and road safety platform
    for Ghana. This API supports live incident reporting, road condition updates,
    and data analytics integration.
  title: Konabra API
  version: "1.0"
paths:
  /account/change:
    post:
      consumes:
      - application/json
      parameters:
      - description: Account change details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/services.ChangeAccountForm'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Send account change verification code (email or phone)
      tags:
      - Account
  /account/change/complete:
    post:
      consumes:
      - application/json
      parameters:
      - description: Account change completion details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/services.CompleteChangeAccountForm'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Complete account change (email or phone)
      tags:
      - Account
  /account/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Account creation details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/services.CreateAccountForm'
      produces:
      - application/json
      responses: {}
      summary: Create a new user account
      tags:
      - Account
  /account/current:
    delete:
      consumes:
      - application/json
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Delete the current user account
      tags:
      - Account
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get current user account
      tags:
      - Account
  /account/password/change:
    post:
      consumes:
      - application/json
      parameters:
      - description: Password change details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/services.ChangePasswordForm'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Change the current password
      tags:
      - Account
  /account/password/reset:
    post:
      consumes:
      - application/json
      parameters:
      - description: Password reset details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/services.ResetPasswordForm'
      produces:
      - application/json
      responses: {}
      summary: Send password reset code
      tags:
      - Account
  /account/password/reset/complete:
    post:
      consumes:
      - application/json
      parameters:
      - description: Password reset completion details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/services.CompleteResetPasswordForm'
      produces:
      - application/json
      responses: {}
      summary: Complete password reset
      tags:
      - Account
  /account/signin:
    post:
      consumes:
      - application/json
      parameters:
      - description: Sign-in credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/services.SignInForm'
      produces:
      - application/json
      responses: {}
      summary: Sign in to an existing account
      tags:
      - Account
  /account/signin/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refresh token details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/services.SignInWithRefreshTokenForm'
      produces:
      - application/json
      responses: {}
      summary: Sign in using a refresh token
      tags:
      - Account
  /account/signout:
    post:
      consumes:
      - application/json
      parameters:
      - description: Sign-out request details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/services.SignOutForm'
      produces:
      - application/json
      responses: {}
      summary: Sign out of the current account
      tags:
      - Account
  /account/verify:
    post:
      consumes:
      - application/json
      parameters:
      - description: Verification details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/services.VerifyAccountForm'
      produces:
      - application/json
      responses: {}
      summary: Send account verification code (email or phone)
      tags:
      - Account
  /account/verify/complete:
    post:
      consumes:
      - application/json
      parameters:
      - description: Verification completion details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/services.CompleteVerifyAccountForm'
      produces:
      - application/json
      responses: {}
      summary: Complete account verification (email or phone)
      tags:
      - Account
  /categories:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: offset
        type: integer
      - in: query
        name: order
        type: string
      - in: query
        name: search
        type: string
      - in: query
        name: sort
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get paginated categories
      tags:
      - Categories
  /categories/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Category Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get category by Id
      tags:
      - Categories
  /categories/all:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: order
        type: string
      - in: query
        name: search
        type: string
      - in: query
        name: sort
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get all categories
      tags:
      - Categories
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"